refactor: separate navigation logic & screens from other components
