+++ dev-client/src/components/dataInputs/SoilScreen.tsx
@@ -21,10 +21,8 @@ import {useTranslation} from 'react-i18next';
 import {Icon, IconButton} from 'terraso-mobile-client/components/common/Icons';
 import {AddIntervalModal} from 'terraso-mobile-client/components/dataInputs/AddIntervalModal';
 import {BottomSheetModalProvider} from '@gorhom/bottom-sheet';
-import {
-  BottomSheetModal,
-  Modal,
-} from 'terraso-mobile-client/components/common/Modal';
+import {Modal} from 'terraso-mobile-client/components/modals/Modal';
+import {BottomSheetModal} from 'terraso-mobile-client/components/modals/BottomSheetModal';
 import {EditIntervalModal} from 'terraso-mobile-client/components/dataInputs/EditIntervalModal';
 import {useMemo, useCallback} from 'react';
 import {
