+++ dev-client/src/components/sites/ProjectInstructionsButton.tsx
@@ -18,7 +18,7 @@
 import {Text, HStack, Box, Button} from 'native-base';
 import {useTranslation} from 'react-i18next';
 import {Icon} from 'terraso-mobile-client/components/common/Icons';
-import {useNavigation} from 'terraso-mobile-client/screens/useNavigation';
+import {useNavigation} from 'terraso-mobile-client/navigation/useNavigation';
 import {useCallback} from 'react';
 import {Project} from 'terraso-client-shared/project/projectSlice';
