+++ dev-client/src/context/HeaderHeightContext.tsx
@@ -1,0 +1,4 @@
+import {createContext, useContext} from 'react';
+
+export const HeaderHeightContext = createContext<number | undefined>(undefined);
+export const useHeaderHeight = () => useContext(HeaderHeightContext);
