+++ dev-client/src/native/checkAndroidPermissions.ts
