+++ dev-client/src/navigation/buttons/AppBarIconButton.tsx
@@ -1,0 +1,8 @@
+import {
+  IconButton,
+  IconButtonProps,
+} from 'terraso-mobile-client/components/common/Icons';
+
+export const AppBarIconButton = (props: IconButtonProps) => (
+  <IconButton size="md" _icon={{color: 'primary.contrast'}} {...props} />
+);
