+++ dev-client/src/navigation/buttons/BottomNavIconButton.tsx
@@ -1,0 +1,8 @@
+import {
+  IconButton,
+  IconButtonProps,
+} from 'terraso-mobile-client/components/common/Icons';
+
+export const BottomNavIconButton = (
+  props: IconButtonProps & {label: string},
+) => <IconButton pb={0} _icon={{color: 'primary.contrast'}} {...props} />;
