+++ dev-client/src/navigation/buttons/ScreenCloseButton.tsx
@@ -1,0 +1,3 @@
+import {ScreenBackButton} from './ScreenBackButton';
+
+export const ScreenCloseButton = () => <ScreenBackButton icon="close" />;
