+++ dev-client/src/screens/AddSiteNoteScreen.tsx
@@ -23,7 +23,7 @@ import {ScreenFormWrapper} from 'terraso-mobile-client/components/common/ScreenF
 import {useTranslation} from 'react-i18next';
 import {SiteNoteAddMutationInput} from 'terraso-client-shared/graphqlSchema/graphql';
 import {addSiteNote} from 'terraso-client-shared/site/siteSlice';
-import {useNavigation} from 'terraso-mobile-client/screens/useNavigation';
+import {useNavigation} from 'terraso-mobile-client/navigation/useNavigation';
 import {Keyboard} from 'react-native';
 
 type Props = {
