+++ dev-client/src/screens/AddUserToProjectScreen/index.tsx
@@ -18,18 +18,16 @@
 import {Box, Button, HStack, Text} from 'native-base';
 import {FreeformTextInput} from 'terraso-mobile-client/components/common/FreeformTextInput';
 import {useTranslation} from 'react-i18next';
-import {
-  AppBar,
-  ScreenScaffold,
-} from 'terraso-mobile-client/screens/ScreenScaffold';
+import {ScreenScaffold} from 'terraso-mobile-client/screens/ScreenScaffold';
+import {AppBar} from 'terraso-mobile-client/navigation/AppBar';
 import {useCallback, useMemo, useState} from 'react';
 import {UserRole} from 'terraso-client-shared/graphqlSchema/graphql';
 import {checkUserInProject} from 'terraso-client-shared/account/accountService';
 import MembershipControlList, {
   UserWithRole,
-} from 'terraso-mobile-client/components/projects/AddUserToProjectScreen/MembershipControlList';
+} from 'terraso-mobile-client/screens/AddUserToProjectScreen/MembershipControlList';
 import {addUserToProject} from 'terraso-client-shared/project/projectSlice';
-import {useNavigation} from 'terraso-mobile-client/screens/useNavigation';
+import {useNavigation} from 'terraso-mobile-client/navigation/useNavigation';
 import {useDispatch, useSelector} from 'terraso-mobile-client/model/store';
 import {useKeyboardOpen} from 'terraso-mobile-client/hooks/index.hooks';
