--- dev-client/src/screens/AppScaffold.tsx
@@ -1,151 +1,0 @@
-/*
- * Copyright © 2023 Technology Matters
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program. If not, see https://www.gnu.org/licenses/.
- */
-
-import {
-  NativeStackNavigationOptions,
-  NativeStackScreenProps,
-  createNativeStackNavigator,
-} from '@react-navigation/native-stack';
-import {useDispatch, useSelector} from 'terraso-mobile-client/model/store';
-import React, {useEffect} from 'react';
-import {
-  setHasAccessTokenAsync,
-  fetchUser,
-} from 'terraso-client-shared/account/accountSlice';
-import {LoginScreen} from 'terraso-mobile-client/screens/LoginScreen';
-import {ProjectListScreen} from 'terraso-mobile-client/screens/ProjectListScreen';
-import {ProjectViewScreen} from 'terraso-mobile-client/screens/ProjectViewScreen';
-import {CreateProjectScreen} from 'terraso-mobile-client/screens/CreateProjectScreen';
-import {HomeScreen} from 'terraso-mobile-client/screens/HomeScreen';
-import {SiteTransferProjectScreen} from 'terraso-mobile-client/screens/SiteTransferProjectScreen';
-import {CreateSiteScreen} from 'terraso-mobile-client/screens/CreateSiteScreen';
-import {AddSiteNoteScreen} from 'terraso-mobile-client/components/siteNotes/AddSiteNoteScreen';
-import {EditSiteNoteScreen} from 'terraso-mobile-client/components/siteNotes/EditSiteNoteScreen';
-import {ReadNoteScreen} from 'terraso-mobile-client/components/siteNotes/ReadNoteScreen';
-import {EditProjectInstructionsScreen} from 'terraso-mobile-client/components/projects/EditProjectInstructionsScreen';
-import {LocationDashboardScreen} from 'terraso-mobile-client/screens/LocationDashboardScreen';
-import {SiteSettingsScreen} from 'terraso-mobile-client/components/sites/SiteSettingsScreen';
-import {SiteTeamSettingsScreen} from 'terraso-mobile-client/components/sites/SiteTeamSettings';
-import {Location, locationManager} from '@rnmapbox/maps';
-import {updateLocation} from 'terraso-mobile-client/model/map/mapSlice';
-import {USER_DISPLACEMENT_MIN_DISTANCE_M} from 'terraso-mobile-client/constants';
-import {AddUserToProjectScreen} from 'terraso-mobile-client/components/projects/AddUserToProjectScreen';
-import {ManageTeamMemberScreen} from 'terraso-mobile-client/components/projects/ManageTeamMemberScreen';
-
-type UnknownToUndefined<T extends unknown> = unknown extends T ? undefined : T;
-export type ScreenDefinitions = Record<string, React.FC<any>>;
-export type ModalScreenDefinitions = Record<string, React.FC<any>>;
-export type ParamList<T extends ScreenDefinitions> = {
-  [K in keyof T]: UnknownToUndefined<React.ComponentProps<T[K]>>;
-};
-
-const screenDefinitions = {
-  LOGIN: LoginScreen,
-  PROJECT_LIST: ProjectListScreen,
-  PROJECT_VIEW: ProjectViewScreen,
-  HOME: HomeScreen,
-  CREATE_PROJECT: CreateProjectScreen,
-  SITE_TRANSFER_PROJECT: SiteTransferProjectScreen,
-  CREATE_SITE: CreateSiteScreen,
-  LOCATION_DASHBOARD: LocationDashboardScreen,
-  SITE_SETTINGS: SiteSettingsScreen,
-  SITE_TEAM_SETTINGS: SiteTeamSettingsScreen,
-  ADD_USER_PROJECT: AddUserToProjectScreen,
-  MANAGE_TEAM_MEMBER: ManageTeamMemberScreen,
-} satisfies ScreenDefinitions;
-
-const modalScreenDefinitions = {
-  ADD_SITE_NOTE: AddSiteNoteScreen,
-  EDIT_SITE_NOTE: EditSiteNoteScreen,
-  EDIT_PROJECT_INSTRUCTIONS: EditProjectInstructionsScreen,
-  READ_NOTE: ReadNoteScreen,
-} satisfies ModalScreenDefinitions;
-
-const combinedScreenDefinitions = {
-  ...screenDefinitions,
-  ...modalScreenDefinitions,
-} satisfies ScreenDefinitions;
-
-export type RootStackParamList = ParamList<typeof combinedScreenDefinitions>;
-export type ScreenName = keyof RootStackParamList;
-
-const RootStack = createNativeStackNavigator<RootStackParamList>();
-
-export type RootStackScreenProps = NativeStackScreenProps<RootStackParamList>;
-
-const screens = Object.entries(screenDefinitions).map(([name, Screen]) => (
-  <RootStack.Screen
-    name={name as ScreenName}
-    key={name}
-    children={props => <Screen {...((props.route.params ?? {}) as any)} />}
-  />
-));
-
-const modalScreens = Object.entries(modalScreenDefinitions).map(
-  ([name, Screen]) => (
-    <RootStack.Screen
-      name={name as ScreenName}
-      key={name}
-      children={props => <Screen {...((props.route.params ?? {}) as any)} />}
-    />
-  ),
-);
-
-const defaultScreenOptions: NativeStackNavigationOptions = {headerShown: false};
-
-export function AppScaffold() {
-  const dispatch = useDispatch();
-  const hasToken = useSelector(state => state.account.hasToken);
-  const currentUser = useSelector(state => state.account.currentUser.data);
-  const loggedIn = useSelector(
-    state => state.account.currentUser.data !== null,
-  );
-
-  useEffect(() => {
-    if (!hasToken) {
-      dispatch(setHasAccessTokenAsync());
-    }
-    if (hasToken && currentUser === null) {
-      dispatch(fetchUser());
-    }
-  }, [hasToken, currentUser, dispatch]);
-
-  useEffect(() => {
-    const listener = ({coords}: Location) =>
-      dispatch(
-        updateLocation({coords: coords, accuracyM: coords.accuracy ?? null}),
-      );
-    locationManager.setMinDisplacement(USER_DISPLACEMENT_MIN_DISTANCE_M);
-    locationManager.addListener(listener);
-    return () => locationManager.removeListener(listener);
-  }, [dispatch]);
-
-  return (
-    <RootStack.Navigator initialRouteName={loggedIn ? 'HOME' : 'LOGIN'}>
-      <RootStack.Group screenOptions={defaultScreenOptions}>
-        {screens}
-      </RootStack.Group>
-      <RootStack.Group
-        screenOptions={{
-          headerShown: false,
-          animation: 'slide_from_bottom',
-        }}>
-        {modalScreens}
-      </RootStack.Group>
-    </RootStack.Navigator>
-  );
-}
